Fix constraints for RGB LEDs
Looking at the schematic, pins 39, 40, and 41 are connected to RGB0,
RGB1 and RGB2 respectively, which are then connected to the blue, green
and red LEDs respectively within the RGB LED. This fixes all the
constraints in this repository to respect that.

`SB_RGBA_DRV` seems to ignore what you connect its outputs to and
output to the `LED*` pins no matter what, so it needs to have its
arguments rearranged as well.